refactor: 計算ロジックを専用モジュールに抽出
- calculateDividends を dividendCalculations.ts に抽出
- calculateDailyData, calculateDomesticStock を domesticStockCalculations.ts に抽出
- ページコンポーネントの責務を軽減
