+++ frontend/src/features/receipt/calculations/index.ts
@@ -1,0 +1,2 @@
+export { calculateDividends } from './dividendCalculations';
+export { calculateDailyData, calculateDomesticStock } from './domesticStockCalculations';
