+++ frontend/src/pages/Receipt/DomesticStock.tsx
@@ -2,16 +2,11 @@ import { ReceiptTemplate } from '@/components/templates/ReceiptTemplate';
 import { ReceiptHeader } from '@/components/molecules/ReceiptHeader/ReceiptHeader';
 import { ReceiptTable } from '@/components/organisms/ReceiptTable/ReceiptTable';
 import React, { useState } from 'react';
-import {
-    DomesticStockData,
-    DomesticStockCalculations,
-    DomesticStockSummary
-} from '@/lib/interfaces/domesticStock';
+import { DomesticStockData } from '@/lib/interfaces/domesticStock';
 import { TableColumnConfig, SummaryColumnConfig } from '@/lib/interfaces/receipt';
 import { createSearchOptions } from '@/lib/utils/dataTransformer';
 import {
     formatJPDate,
-    TAX_RATE,
     formatCurrency,
     formatNumber,
     createISODateKey
@@ -25,58 +20,7 @@ import {
     matchesAmounts
 } from '@/lib/utils/searchUtils';
 import { parseDomesticStockCsvItem, sortDomesticStockByTradeDate } from '@/features/receipt/parsers';
-
-// 日次データ集計
-const calculateDailyData = (domesticStockData: DomesticStockData[]): DomesticStockSummary[] => {
-    // データを日付ごとにグループ化
-    const dailyGroupMap = new Map<string, DomesticStockData[]>();
-
-    domesticStockData.forEach(item => {
-        const dateKey = createISODateKey(item.trade_date);
-        if (!dailyGroupMap.has(dateKey)) {
-            dailyGroupMap.set(dateKey, []);
-        }
-        dailyGroupMap.get(dateKey)?.push(item);
-    });
-
-    // 日次データの集計
-    return Array.from(dailyGroupMap.entries()).map(([dateKey, items]) => {
-        // 特定口座とNISA口座の集計を分離
-        const dailyTotals = items.reduce((acc, item) => {
-            const isSpecificAccount = item.account.includes('特定');
-            return {
-                specificTotal: acc.specificTotal + (isSpecificAccount ? item.realized_profit_and_loss : 0),
-                nisaTotal: acc.nisaTotal + (!isSpecificAccount ? item.realized_profit_and_loss : 0),
-                amount: acc.amount + item.proceeds
-            };
-        }, { specificTotal: 0, nisaTotal: 0, amount: 0 });
-
-        // 実現損益の計算
-        const totalRealizedPnL = dailyTotals.specificTotal + dailyTotals.nisaTotal;
-        const tax = Math.floor(Math.max(0, dailyTotals.specificTotal) * TAX_RATE);
-        const totalRealizedPnLAfterTax = dailyTotals.specificTotal - tax + dailyTotals.nisaTotal;
-
-        return {
-            filter: dateKey,
-            total_realized_profit_and_loss: totalRealizedPnL,
-            total_taxes: tax,
-            total_realized_profit_and_loss_after_tax: totalRealizedPnLAfterTax,
-        };
-    }).sort((a, b) => a.filter.localeCompare(b.filter));
-};
-
-// 全体集計関数
-const calculateDomesticStock = (dailyData: DomesticStockSummary[]): DomesticStockCalculations => {
-    return dailyData.reduce((acc, item) => ({
-        total_realized_profit_and_loss: acc.total_realized_profit_and_loss + item.total_realized_profit_and_loss,
-        total_taxes: acc.total_taxes + item.total_taxes,
-        total_realized_profit_and_loss_after_tax: acc.total_realized_profit_and_loss_after_tax + item.total_realized_profit_and_loss_after_tax
-    }), {
-        total_realized_profit_and_loss: 0,
-        total_taxes: 0,
-        total_realized_profit_and_loss_after_tax: 0,
-    });
-};
+import { calculateDailyData, calculateDomesticStock } from '@/features/receipt/calculations';
 
 interface DomesticStockProps {
     csvData: Record<string, unknown>[];
